LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-canvas-element/security.pattern.image.strokeStyle.sub-expected.txt

CONSOLE MESSAGE: line 27: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
security.pattern.image.strokeStyle.sub
Setting strokeStyle to a pattern of a different-origin image makes the canvas origin-unclean
Actual output:

PASS Setting strokeStyle to a pattern of a different-origin image makes the canvas origin-unclean 
